Added a few things/fixed a few things
Added messages for the Request Buffs button. Players can now choose what message they would like to use instead of the hardcoded one.
Add the ability to change the fonts, color, and size for the Conquest and Honor tabs.
Also made it to where honor points and conquest update live while the player is playing.