Fixed a few things
Buttons now look like they are actually being clicked (depressed). Also made it to if a button is clicked it will make a sound. The sound is the same one as if you are in game and open the game menu. Also got rid of the red that was under the button whenever you clicked it.