Fixed a few things
Removed UIPanelButtonTemplate, it was causing issues with the button border as well as having the color red as the buttons. My textures just masked the red until the button was clicked and it would show red. That is now fixed as well as the button borders. They now look a lot better.