Made a small change
The only time the map will be able to be sized will be in an PVP environment. So lets say you enter a BG and you resize the map the, map will resize like you want. But, once you leave the BG the map will return to its normal default size. When you enter the next BG the map will resize to what you had previously.